+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Breaking changes:
 * Update ruma-identifiers to 0.15.1
 * Change timestamps, including `origin_server_rs` from `UInt` to `SystemTime`
 * Change all usages of `HashMap` to `BTreeMap`
+  * To support this, `EventType` now implements `PartialOrd` and `Ord`
 
 # 0.18.0